% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{adjust_nas}
\alias{adjust_nas}
\title{Impute Missing Values}
\usage{
adjust_nas(data, na_decision, neutrals)
}
\arguments{
\item{data}{the raw data}

\item{na_decision}{a determination by the user if NAs should be
removed or replaced}

\item{neutrals}{the average integer value between the maximum and the minimum
input values in each question}
}
\description{
Handles \code{NA} values based on user preference, specifically
allowing replacement with the calculated scale midpoint.
}
