% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{bias_handling}
\alias{bias_handling}
\title{Handle Response Bias and Low-Quality Entries}
\usage{
bias_handling(data, neutrals, col_mins, col_maxs)
}
\arguments{
\item{data}{the raw data}

\item{neutrals}{the average integer value between the maximum and the minimum
input values in each question}

\item{col_mins}{the lowest answered value in each question}

\item{col_maxs}{the largest answered value in each question}
}
\description{
Filters out "low-effort" respondents, including:
\itemize{
\item \strong{Fence-sitters}: Respondents who only chose the
neutral midpoint.
\item \strong{Straight-liners}: Respondents who only chose
the scale minimum or maximum.
\item \strong{Identical responses}: Respondents who gave the
same value for every question.
}
}
