% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbachs_alpha.R
\name{c_alpha_calc}
\alias{c_alpha_calc}
\title{c_alpha_calc}
\usage{
c_alpha_calc(data)
}
\arguments{
\item{data}{a dataframe where each column is a survey question (item)
and each row is a response.}
}
\value{
a numeric value. cronbach's alpha for the provided data.
}
\description{
function for calculating cronbach's alpha for a dataset
}
\details{
this function handles the math for a prepared and filtered dataset.
See \code{\link[=cronbachs_alpha]{cronbachs_alpha()}} for the process of the analysis.
}
