% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{converting_to_percentage}
\alias{converting_to_percentage}
\title{Frequency to Percentage Calculation}
\usage{
converting_to_percentage(col)
}
\arguments{
\item{col}{the responses for each question}
}
\description{
Calculates the frequency of each response level within a
specific column, relative to the observed scale range, and returns
rounded percentages.
}
