% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbachs_alpha.R
\name{cronbachs_alpha}
\alias{cronbachs_alpha}
\title{cronbachs_alpha}
\usage{
cronbachs_alpha(data, groups)
}
\arguments{
\item{data}{a dataframe where each column is a likert survey question (item)
and each row is a response.}

\item{groups}{a numeric vector specifying groups of questions.
Cronbach's alpha information will be calculated separately for each
group. The vector must be the same length as the number of items
(columns) in the dataframe. Ex. c(1, 1, 2, 2) i.e questions 1-2 are in
a group and questions 2-4 are in a group. Groups should be numbered
1, 2, 3, ... , n.}
}
\value{
a list containing lists of Cronbach's alpha information for each
group. Each group sub-list contains the Cronbach's alpha for the group
as well as the Cronbach's alpha when each item (question) is left out.
}
\description{
used to find Cronbach's alpha information for a dataset.
}
