% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa.R
\name{efa}
\alias{efa}
\title{Carry out exploratory factor analysis}
\usage{
efa(data, n)
}
\arguments{
\item{data}{A dataset in the form of a data.frame that will be used for the
analysis}

\item{n}{An optional integer argument indicating how many factors should be
used in the analysis}
}
\value{
A list containing smaller lists of pre-EFA diagnostics and EFA
results, which both contain several measures
}
\description{
\code{efa()} carries out exploratory factor analysis on a dataset,
returning pre-EFA diagnostics as well as various EFA results
}
