% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{general_cleaning}
\alias{general_cleaning}
\title{Initial Data Filtering and Question Extraction}
\usage{
general_cleaning(data)
}
\arguments{
\item{data}{the raw data}
}
\description{
Identifies numeric columns that follow a Likert format (integers
between 1 and 11) and renames them for internal processing.
}
