% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{ipsatize}
\alias{ipsatize}
\title{Person-Mean Centering (Ipsatization)}
\usage{
ipsatize(data, ipsatize_decision)
}
\arguments{
\item{data}{the raw data}

\item{ipsatize_decision}{a determination by the user if this data is wanted}
}
\description{
Subtracts the respondent's average score from
each of their individual responses to control for individual
response styles (e.g., general tendency to agree or disagree).
}
