% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likertr.R
\name{likertr}
\alias{likertr}
\title{Likertr Wrapper Function}
\usage{
likertr(
  data,
  n_fact,
  na_drop = FALSE,
  ipsatize_decision = FALSE,
  small_n_drop = FALSE,
  groups = numeric(0),
  factor_inference = NA,
  inference_vars = NA,
  inference_vars2 = NA,
  flip = FALSE,
  plot = FALSE
)
}
\arguments{
\item{data}{a dataframe where each column is a likert survey question (item)
and each row is a response.}

\item{n_fact}{Integer. Optional argument for the number of factors to be used
in the Exploratory Factor Analysis (EFA). If not given, this argument will
be decided using parallel analysis.}

\item{na_drop}{Character. Strategy for handling \code{NA} values; if "neutral",
missing values are replaced with the scale midpoint.}

\item{ipsatize_decision}{Logical. If \code{TRUE}, returns a version of the data
centered by respondent (person-mean centering).}

\item{small_n_drop}{Character. If not "nothing", questions/groups with
fewer than 20 responses are dropped.}

\item{groups}{a numeric vector specifying groups of questions.
Cronbach's alpha information will be calculated separately for each
group. The vector must be the same length as the number of items
(columns) in the dataframe. Ex. c(1, 1, 2, 2) i.e questions 1-2 are in
a group and questions 2-4 are in a group. Groups should be numbered
1, 2, 3, ... , n.}

\item{factor_inference}{column index referencing a factor variable to split
data on for inference.}

\item{flip}{If flip is TRUE, then items are automatically flipped to have
positive correlations on the general factor (recommended to do this
manually before)}

\item{plot}{Whether or not to call omega.diagram}
}
\description{
Likertr Wrapper Function
}
