% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdonalds_omega.R
\name{mcdonalds_omega}
\alias{mcdonalds_omega}
\title{mcdonalds_omega}
\usage{
mcdonalds_omega(data, fm, flip, plot, rotate)
}
\description{
function for calculating McDonald's Omega for likert survey data

DOCUMENTATION :)
}
