% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{noting_small_n}
\alias{noting_small_n}
\title{Sample Size Validation}
\usage{
noting_small_n(data, num_people, small_n_decision)
}
\arguments{
\item{data}{the raw data}

\item{num_people}{the number of people's responses per question}

\item{small_n_decision}{a determination by the user if questions with
low numbers of responses should be dropped.}
}
\description{
Checks if questions meet a minimum threshold (n=20) and
issues warnings or removes data accordingly.
}
