% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.likertr.R
\name{plot.likertr}
\alias{plot.likertr}
\title{Plot Method for Likertr Objects}
\usage{
\method{plot}{likertr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{likertr}, typically the output from
the \code{preparation} function}

\item{...}{Additional arguments passed to the underlying plotting functions.}
}
\description{
A S3 method for objects of class \code{likertr}.
This function automatically generates a set of visualizations.
}
