% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{readme_data}
\alias{readme_data}
\title{Readme Example Data}
\format{
A data frame with X rows and Y variables
}
\usage{
readme_data
}
\description{
Data used for the README examples.
}
\keyword{datasets}
