% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rii.R
\name{rii}
\alias{rii}
\title{rii}
\usage{
rii(data, max_val)
}
\arguments{
\item{data}{a dataframe where each column is a likert survey question (item)
and each row is a response.}

\item{max_val}{a numeric vector with length equal to the number of items in
the dataframe. Specifies the point scale of each question by providing
the max value answerable (eg. if a question is on a five point scale
the value provided in max_val should be five).}
}
\value{
a numeric vector with the relative importance index for each
question.
}
\description{
used to calculate the relative importance index for each item
(question) in a likert survey.
}
