% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{split_question}
\alias{split_question}
\title{Convert Raw Data to Percentages}
\usage{
split_question(data)
}
\arguments{
\item{data}{the raw data}
}
\description{
Takes the cleaned data frame and splits it into a list of
percentage distributions, one for each question, suitable
for Likert visualization. Calls the helper function
\code{\link[=converting_to_percentage]{converting_to_percentage()}}.
}
