% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.likertr.R
\name{stacked_bar}
\alias{stacked_bar}
\title{Standard Stacked Bar Chart for Likert Data}
\usage{
stacked_bar(perc_by_question, questions)
}
\arguments{
\item{perc_by_question}{A list of numeric vectors containing percentage
distributions for each question.}

\item{questions}{A character vector of labels/questions to display on the
y-axis.}
}
\description{
Generates a horizontal stacked bar chart where each bar represents a question
and the segments represent the percentage distribution of responses. This
visualization uses a color gradient to distinguish between response levels
and places a legend on the left side of the plot.
}
