% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.likertr.R
\name{summary.likertr}
\alias{summary.likertr}
\title{Summarize the results of a likertr object}
\usage{
\method{summary}{likertr}(
  object,
  data_summary = TRUE,
  efa_summary = TRUE,
  reliability_summary = TRUE,
  inference_summary = TRUE,
  ...
)
}
\arguments{
\item{object}{A likertr object for which to generate a summary report}

\item{data_summary}{Logical. Determines whether or not a data summary
is included in the \code{\link[=summary]{summary()}} output.}

\item{efa_summary}{Logical. Determines whether or not an exploratory
factor analysis summary is included in the \code{\link[=summary]{summary()}} output.}

\item{reliability_summary}{Logical. Determines whether or not a
reliability summary is included in the \code{\link[=summary]{summary()}} output.}

\item{inference_summary}{Logical. Determines whether or not an inference
summary is included in the \code{\link[=summary]{summary()}} output.}

\item{...}{All other arguments to be passed to this function.}
}
\description{
\code{summary.likertr()} is a generic summary function for objects of the
likertr class. The function generates a summary report with EFA
diagnostics and results, reliability measures, and inferential
statistics.
}
