% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likertr.R
\name{likertr}
\alias{likertr}
\title{Likertr Wrapper Function}
\usage{
likertr(
  data,
  max_val,
  na_decision = "drop",
  ipsatize_decision = FALSE,
  small_n_decision = "nothing",
  groups = numeric(0),
  std = TRUE,
  empirical = TRUE
)
}
\arguments{
\item{data}{a dataframe where each column is a likert survey question (item)
and each row is a response.}

\item{max_val}{a numeric vector with length equal to the number of items in
the dataframe. Specifies the point scale of each question by providing
the max value answerable (eg. if a question is on a five point scale
the value provided in max_val should be five).}

\item{na_decision}{Character. Strategy for handling \code{NA} values; if "neutral",
missing values are replaced with the scale midpoint.}

\item{ipsatize_decision}{Logical. If \code{TRUE}, returns a version of the data
centered by respondent (person-mean centering).}

\item{small_n_decision}{Character. If not "nothing", questions/groups with
fewer than 20 responses are dropped.}

\item{groups}{a numeric vector specifying groups of questions.
Cronbach's alpha information will be calculated separately for each
group. The vector must be the same length as the number of items
(columns) in the dataframe. Ex. c(1, 1, 2, 2) i.e questions 1-2 are in
a group and questions 2-4 are in a group. Groups should be numbered
1, 2, 3, ... , n.}

\item{empirical}{}
}
\description{
Likertr Wrapper Function
}
