% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likertr.R
\name{likertr}
\alias{likertr}
\title{Likertr Wrapper Function}
\usage{
likertr(
  data,
  max_val,
  na_decision = "drop",
  ipsatize_decision = FALSE,
  small_n_decision = "nothing",
  groups = numeric(0),
  std = TRUE,
  empirical = TRUE
)
}
\arguments{
\item{empirical}{}
}
\description{
Likertr Wrapper Function
}
