% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likertr.R
\name{likertr}
\alias{likertr}
\title{Likertr Wrapper Function}
\usage{
likertr(
  data,
  na_drop = FALSE,
  n_fact,
  ipsatize_decision = FALSE,
  small_n_drop = FALSE,
  groups = numeric(0),
  factor_inference = NA,
  inference_vars = NA,
  inference_vars2 = NA,
  flip = FALSE,
  plot = FALSE
)
}
\arguments{
\item{data}{a dataframe where each column is a likert survey question (item)
and each row is a response.}

\item{na_drop}{Character. Strategy for handling \code{NA} values; if "neutral",
missing values are replaced with the scale midpoint.}

\item{n_fact}{Integer. Optional argument for the number of factors to be used
in the Exploratory Factor Analysis (EFA). If not given, this argument will
be decided using parallel analysis.}

\item{ipsatize_decision}{Logical. If \code{TRUE}, returns a version of the data
centered by respondent (person-mean centering).}

\item{small_n_drop}{Character. If not "nothing", questions/groups with
fewer than 20 responses are dropped.}

\item{groups}{a numeric vector specifying groups of questions.
Cronbach's alpha information will be calculated separately for each
group. The vector must be the same length as the number of items
(columns) in the dataframe. Ex. c(1, 1, 2, 2) i.e questions 1-2 are in
a group and questions 2-4 are in a group. Groups should be numbered
1, 2, 3, ... , n.}

\item{factor_inference}{integer corresponding to a column index
referencing a factor variable to split data on for inference.}

\item{inference_vars}{vector of integers corresponding to column
indicies to use for inferential analysis}
}
\description{
A comprehensive wrapper for psychometric analysis of
Likert-scale survey data. This function automates the pipeline
from data cleaning (including optional ipsatization and NA handling)
to Exploratory Factor Analysis (EFA), reliability testing
(Cronbach's alpha and McDonald's omega), Relative Importance Index
(RII) calculation, and inferential testing.
}
\examples{
\dontrun{
load("data/data.rda")
# Run the wrapper with 2-group reliability check
# Group 1: Items 1-2, Group 2: Items 3-5
results <- likertr(
  data = data,
  na_drop = "neutral",
  n_fact = 1,
  groups = c(1, 1, 2, 2, 2),
  ipsatize_decision = TRUE
)
}

}
