% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{preparation}
\alias{preparation}
\title{Preparation and Cleaning}
\usage{
preparation(
  data,
  na_decision = "drop",
  ipsatize_decision = FALSE,
  small_n_decision = "nothing"
)
}
\arguments{
\item{small_n_decision}{}
}
\description{
Preparation and Cleaning
}
