% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.likertr.R
\name{summary.likertr}
\alias{summary.likertr}
\title{Summarize the results of a likertr object}
\usage{
\method{summary}{likertr}(object, ...)
}
\arguments{
\item{object}{A likertr object for which to generate a summary report}

\item{...}{All other arguments to be passed to this function.}
}
\description{
\code{summary.likertr()} is a generic summary function for objects of the
likertr class. The function generates a summary report with EFA
diagnostics and results, reliability measures, and inferential
statistics.
}
