% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.likertr.R
\name{summary.likertr}
\alias{summary.likertr}
\title{Summarize the results of a likertr object}
\usage{
\method{summary}{likertr}(x, ...)
}
\arguments{
\item{x}{A likertr object for which to generate a summary report}
}
\description{
\code{summary.likertr()} is a generic summary function for objects of the
likertr class. The function generates a summary report with EFA
diagnostics and results, reliability measures, and inferential
statistics.
}
